+++ src/org/conjugateprior/ca/app/CLCategoryCounter.java
@@ -1,72 +1,42 @@
 package org.conjugateprior.ca.app;
 
-import java.io.File;
 import java.nio.charset.Charset;
-import java.util.Locale;
 
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
 import org.apache.commons.cli.GnuParser;
-import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Option;
-import org.apache.commons.cli.Options;
 
-public class CLCategoryCounter {
+public class CLCategoryCounter extends CLApplication {
 
 	protected CategoryCounter counter;
-	
-	protected Options options;
-	protected HelpFormatter helpFormatter;
-	
-	public static String USAGE = "ykcats [-encoding <encoding>] [-locale <locale>] " +
-	   "[-oldmatching] [-output <folder>] [-format <format>] -dictionary <file> " +
-	   "[doc1.txt doc2.txt folder1]";
-	
-	public void printUsageAndOptions(){
-		helpFormatter.printHelp(USAGE, options);
+		
+	public String getUsage() {
+		return "ykcats [-encoding <encoding>] [-locale <locale>] " +
+				   "[-oldmatching] [-output <folder>] [-format <format>] -dictionary <file> " +
+				   "[doc1.txt doc2.txt folder1]";
 	}
 	
 	public CLCategoryCounter(CategoryCounter c) {
+		super();	
+		
 		counter = c;
 		
-		helpFormatter = new HelpFormatter();
-		options = new Options();
+		addOption(getHelpOption(false));
+		addOption(getEncodingOption(false));
+		addOption(getLocaleOption(false));
+		addOption(getOuputFolderOption(true)); // not file
+		addOption(getDictionaryOption(true));
 		
-		Option oHelp = new Option("help", "Show this message, then exit");
-		Option oEncoding = new Option("encoding", true, 
-				"Input file character encoding (default: " + 
-				Charset.defaultCharset().displayName() + ")");
-		oEncoding.setArgName("encoding");
-		
-		Option oOldMatching = new Option("oldmatching", 
+		CLOption oldmatching = new CLOption("oldmatching", 
 				"Use old-style pattern matching");
-		Option oLocale = new Option("locale",  true, 
-				"Locale for input files (default: " + 
-				Locale.getDefault().toString() + ")");
-		oLocale.setArgName("locale");				
-		// required
-		Option oDictionary = new Option("dictionary", true, 
-				"Content analysis dictionary in Yoshikoder ('.ykd'), " +
-		        "Lexicoder ('.lcd'), Wordstat ('.CAT'), LIWC ('.dic'), or VBPro " + 
-				"('.vbpro') format");
-		oDictionary.setArgName("file");
-		oDictionary.setRequired(true);
+		addOption(oldmatching);
 		
-		Option oOutputfile = new Option("output", true, 
-				"Name for an output file");
-		oOutputfile.setArgName("file");
-		
-		Option oFormat = new Option("format", true, 
-				"One of: utf8, text (default: text, which here means " + Charset.defaultCharset().name() + ")");
-		oFormat.setArgName("format");
-		
-		options.addOption(oHelp);
-		options.addOption(oEncoding);
-		options.addOption(oOldMatching);
-		options.addOption(oLocale);
-		options.addOption(oDictionary);
-		options.addOption(oOutputfile);
-		options.addOption(oFormat);
+		Option format = new Option("format", true, 
+			"One of: utf8, text (default: text, which here means " + 
+			Charset.defaultCharset().name() + ")");
+		format.setArgName("format");
+		addOption(format);
 	}
 	
 	public void processLine(String[] args) throws Exception {	
@@ -75,31 +45,23 @@ public class CLCategoryCounter {
 		
 		if (line.hasOption("help")) {
 			printUsageAndOptions();
-			System.exit(0);
 		}
-		if (line.hasOption("locale")){ // if not we get the default locale
+		if (line.hasOption("locale")){
 			try {
 				counter.setLocale(line.getOptionValue("locale"));
 			} catch (Exception ex){
-				throw new Exception(
-						"Could not parse locale argument.\n" + 
-				        "A valid locale consists of a two letter language codes from ISO 639\n" +
-						"optionally connected by an underscore to a two letter country code\n" +
-				        "from ISO 3166.  See also http://en.wikipedia.org/wiki/BCP_47");
+				throw new Exception(getOptionErrorMessage("locale"));
 			}
-		} 
+		}
 		if (line.hasOption("encoding")){ // if not we get the default charset
 			try {
 				counter.setEncoding(line.getOptionValue("encoding"));
 			} catch (Exception ex){
-				throw new Exception("Could not parse file encoding. Error message follows:\n" +
-						ex.getMessage());
+				throw new Exception(getOptionErrorMessage("encoding"));
 			}
-		} 		
-
-		// we always have a dictionary 
-		File dictionaryFile = new File(line.getOptionValue("dictionary"));
-		counter.setDictionary(dictionaryFile);
+		}
+		
+		counter.setDictionary(line.getOptionValue("dictionary"));
 		counter.setUsingOldMatchStrategy(line.hasOption("oldmatching"));
 
 		// and files
@@ -119,15 +81,11 @@ public class CLCategoryCounter {
 			counter.setOutputFolder(line.getOptionValue("output"));
 				
 		counter.processFiles();
+		
+		System.exit(0);
 	}
 	
-
 	public static void main(String[] args) throws Exception {
-		//ykcats -dictionary ~/Dropbox/blogposts/littledict.vbpro -output thing ~/Dropbox/blogposts/speeches
-		//String[] as = new String[]{"-dictionary", "/Users/will/Dropbox/blogposts/littledict.vbpro",
-		//							/*"-output", "/Users/will/Dropbox/blogposts/ykcatesoutputfolder",*/ 
-		//							"-format", "html", "/Users/will/Dropbox/blogposts/speeches"};
-		
 		CategoryCounter cc = new CategoryCounter();
 		CLCategoryCounter c = new CLCategoryCounter(cc);
 		try {
@@ -138,71 +96,3 @@ public class CLCategoryCounter {
 		}
 	}
 }
-
-/*
-if (tOutputfile == null){			
-	SimpleDocumentTokenizer tokenizer = new SimpleDocumentTokenizer(tLocale);
-	// push out in default local encoding
-	for (TreeItem<DCat> titem : categoryNodesInPrintOrder) 
-		System.out.print("," + StringEscapeUtils.escapeCsv(DCat.getPathAsString(titem, ">")));
-	System.out.println("," + 
-		CSVFXCategoryDictionaryCountPrinter.wordCountHeader);
-
-	for (File f : filesToProcess) {
-		IYoshikoderDocument idoc = 
-				new SimpleYoshikoderDocument(f.getName(), 
-					AbstractYoshikoderDocument.getTextFromFile(f, tEncoding),
-					null, tokenizer);
-		String docline = makeLineFromDocument(dict, idoc, oldMatchStrategy);
-		System.out.println(docline);
-	}
-} else {
-	CountPrinter printer = null;
-	if (oldMatchStrategy){
-		printer = new CSVOldStyleCategoryDictionaryCountPrinter(dict, 
-				tOutputfile, "data.csv", filesToProcess, tEncoding, tLocale);
-	} else {
-		printer = new CSVFXCategoryDictionaryCountPrinter(dict, 
-				tOutputfile, "data.csv", filesToProcess, tEncoding, tLocale);
-	}
-	final float maxProg = (float)printer.getMaxProgress();
-	final DecimalFormat df = new DecimalFormat("#.##");
-	PropertyChangeListener listener = new PropertyChangeListener() {
-		@Override
-		public void propertyChange(PropertyChangeEvent evt) {
-			if ("progress".equals(evt.getPropertyName())){
-				int prog = (Integer)evt.getNewValue();
-				System.err.println(df.format((prog / maxProg) * 100) + "% complete");
-			}
-		}
-	};
-	printer.addPropertyChangeListener(listener);		
-	printer.processFiles(false);
-
-	String newl = printer.getNewline();
-	File rme = new File(tOutputfile, printer.getReadmefilename());
-	try (
-			OutputStreamWriter out = new OutputStreamWriter(
-					new FileOutputStream(rme, true), 
-					printer.getOutputCharset());
-			BufferedWriter writer = new BufferedWriter(out);
-			){
-		writer.write(newl);
-		writer.write("Settings:");
-		writer.write(newl + newl);
-		writer.write("File enc:\t" + printer.getOutputCharset());
-		writer.write(newl);
-		//writer.write("Output enc:\t" + 
-		//		(onWindows() ? "windows-1252 ('Latin 1')" : "UTF-8"));
-		//writer.write(newl);
-		writer.write("Dict:\t" + line.getOptionValue("dictionary") + " (source file)");
-		writer.write(newl);
-		writer.write("Matching:\t" + (oldMatchStrategy ? "old" : "new"));
-		writer.write(newl);
-		writer.write("Line endings:\t \\n (Unix style)");
-		//writer.write(newl);
-
-	}
-}
-*/
-