+++ src/org/conjugateprior/ca/app/Description.java
@@ -14,35 +14,51 @@ import org.conjugateprior.ca.AbstractYoshikoderDocument;
 import org.conjugateprior.ca.IYoshikoderDocument;
 import org.conjugateprior.ca.SimpleDocumentTokenizer;
 import org.conjugateprior.ca.SimpleYoshikoderDocument;
+import org.conjugateprior.ca.app.AbstractCounter.OutputFormat;
 
 public class Description extends AbstractCounter {
 
 	protected Set<String> vocab = new HashSet<String>();
 	protected String descriptionFilename = "description.csv";
 	
+	protected File outputFile;
+	
+	public File getOutputFile() {
+		return outputFile;
+	}
+	
+	public void setOutputFile(File outputFile) {
+		this.outputFile = outputFile;
+	}
+	
 	static class Desc {
 		int wordCount;
 		int tokenCount; 
 		int hapaxes;
 		double propOfVocab;
+		int sentences;
 		
 		static private DecimalFormat df = new DecimalFormat("#.###");
 		
-		public Desc(int wc, int tk, int hap) {
+		public Desc(int wc, int tk, int hap, int sent) {
 			wordCount = wc;
 			tokenCount = tk;
 			hapaxes = hap;
+			sentences = sent;
 		}
 		
 		public String toString(String sep) {
 			return "" + wordCount + sep + tokenCount + sep + hapaxes + sep +
 					df.format(propOfVocab) + sep + 
-					df.format(((double)tokenCount)/wordCount) ;
+					df.format(((double)tokenCount)/wordCount)  + sep + 
+					sentences;
 		}
 
+		// TODO make sentence stuff print
 		static public String getHeader(String sep) {
 			return "Document" + sep + "TokenCount" + sep + "TypeCount" + 
-					sep + "HapaxCount" + sep + "PropVocabUsed" + sep + "TypeTokenRatio";
+					sep + "HapaxCount" + sep + "PropVocabUsed" + 
+					sep + "TypeTokenRatio" + sep + "SentenceCount";
 		}
 		
 	}
@@ -67,18 +83,15 @@ public class Description extends AbstractCounter {
 				if (entry.getValue() == 1) hap += 1;
 			vocab.addAll(map.keySet());
 			
-			Desc d = new Desc(idoc.getDocumentLength(), map.size(), hap);
+			Desc d = new Desc(idoc.getDocumentLength(), map.size(), 
+					hap, idoc.getSentenceCount());
+			d.sentences = idoc.getSentenceCount();
 			docToDesc.put(f.getName(), d); // in on the filename 
 		}
-		int wcAll = 0;
-		int tkAll = 0;
-		//Desc all = new Desc(wc, tk, hap);
 		
 		double vocabSize = vocab.size();
-		if (outputFolder != null){
-			FileUtils.forceMkdir(outputFolder);
-			File ff = new File(outputFolder, descriptionFilename);	
-			try (BufferedWriter writer = getBufferedWriter(ff)){
+		if (outputFile != null){	
+			try (BufferedWriter writer = getBufferedWriter(outputFile)){
 				writer.write(Desc.getHeader(","));
 				writer.newLine();
 				
@@ -89,7 +102,7 @@ public class Description extends AbstractCounter {
 					writer.write("," + d.toString(","));
 					writer.newLine();
 				}
-				writer.flush(); // do we need this really?			
+				writer.flush(); // do we need this really?
 			}		
 		} else {
 			try (BufferedWriter writer = getBufferedWriter()){